% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfun.R
\name{gibbsR}
\alias{gibbsR}
\title{A Gibbs sampler using R}
\usage{
gibbsR(N, a, b, n)
}
\arguments{
\item{N}{the number of samples}

\item{a}{a param of beta distribution}

\item{b}{another param of beta distribution}

\item{n}{sample size}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
rnR <- gibbsR(100,2,3,10)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
