% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmax}
\alias{lmax}
\title{Compute the maximum l2 norm for each column of the matrix}
\usage{
lmax(x)
}
\arguments{
\item{x}{the object matrix}
}
\value{
the maximum l2 norm for each column of the matrix
}
\description{
Compute the maximum l2 norm for each column of the matrix
}
\examples{
\dontrun{
x <- matrix(c(1, 2, 4, 9), 2, 2)
print(lmax(x))
}
}
