\name{SA23204168-package}
\alias{SA23204168-package}
\alias{SA23204168}
\docType{package}
\title{
  Final Project of Statistical Computing
}
\description{
  Final Project of Statistical Computing that includes all my homework answers and some functions for my research project.
}
\author{
Heng Liu
Maintainer: Heng Liu <liuheng2001@mail.ustc.edu.cn>
}
\keyword{ package }
