% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfun.R
\name{sdr}
\alias{sdr}
\title{sparse sufficient dimension reduction}
\usage{
sdr(
  X,
  y,
  m,
  lambda,
  K,
  nu = 0.01,
  alpha = 0.01,
  epsilon = 0.01,
  maxiter = 10,
  trace = FALSE,
  initPi = NULL,
  initH = NULL
)
}
\arguments{
\item{X}{the n by p predictor, p>n}

\item{y}{the n by q response}

\item{m}{minibatct size}

\item{lambda}{tuning parameter}

\item{K}{Structural dimension}

\item{initPi}{initial value}

\item{initH}{initial value}
}
\value{
p by K basis matrix of sufficient dimension reduction subspace
}
\description{
a function estimates the sparse sufficient subspace
}
\examples{
\dontrun{
rnR <- gibbsR(100,2,3,10)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
